symbols/si: <TLDE> should print cedilla, diaeresis and have only two levels
During a recent discussion about the upcoming new slovenian layout it was
pointed out that the existing is buggy when it comes to the "tilda key".
Instead of printing the cedilla (not dead) and diaresis on the second
level, it prints grave, tilde, notsign and notsign. This is due to historic
reasons, since the various ex-yugoslavian variants were the same.

Signed-off-by: Jaka Kranjc <lynxlynxlynx@sourcemage.org>