Added Crimean Tatar (Crimean Turkish) layouts again, based on layouts released with Ubuntu 9.04
Minor tweaks have been made in comparison to code that was unjustifiably and abusively reverted by Sergey in February. In particular, crh.po is not checked in this time, because it was one of Sergey's reasons for reverting, so until it's pulled from TP, something like `cp po/tr.po po/crh.po` is required prior to building from source.