Added Crimean Tatar (Crimean Turkish) resource bundle as well, b.fd.o#19730
Added Crimean Tatar (Crimean Turkish) resource bundle as well
(as was done originally on January 30),
based on Peter Hutterer's feedback regarding build process, even though
check-in of .po was mentioned by Sergey as one of the reasons for reverting
on February 5. It does in fact make sense to be able to make a build with
the new localization prior to code freeze, so I'm going with my original
understanding, which is apparently shared by Peter Hutterer.