Add Euro and New Shekel sign to israeli layout.
According to [1], the standard layout requires the shekel sign on AltGr + a,
although Windows has it on AltGr + 4. The latter is already in use for some
other character, so let's put it where the standard requires it to be.
Euro sign goes on AltGr + e.

This requires the inclusion of level3(ralt_switch) to get the IsoLevel3.

[1] http://en.wikipedia.org/wiki/Shekel_sign

Fedora Bug https://bugzilla.redhat.com/show_bug.cgi?id=542879

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>