Remove Pointer_EnableKeys from default layout, move to option instead.
Pointer_EnableKeys allows the use of the keypad to move the pointer around.
This is a rarely used feature but can be enabled by default with
Shift+NumLock. No desktop or keyboard has a visualisation that the feature
is enabled, leading to user confusion when erroneous mouse clicks or
movements are emitted.

Remove Pointer_EnableKeys from the default keymap and move it to option
"keypad:pointerkeys" instead.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>