assign action for keysym Shift_L by default
The keysym action was up until now not explicitly defined.
Shift did always work by binding the whole key to a real modifier.
But you don't always want to (as in shift:both_capslock)
For some reason assigning an action to Shift_L prevents the <LFSH>
key from getting bound to the Lock modifier when using
shift:both_capslock.

This fixes the misbehavior of shift not shifting due to multiple
mod-mapping (Shift and Lock) when using shift:both_capslock.