Fix henkan key on jp106 keyboard with inet media keys
When support for inet keys was added to the keyboard driver
back in the early 2000 none of the developers thought of
Japanese 106 key keybards which have two extra keys generating
the scan codes 0x79 and 0x7d.
So for keys prefixed with the scan code 0x60 which were not
remapped to anything else the line:
 *scanCode += 0x78;
was added.
Thus keys generating the scan codes 0xe0 0x01 and 0xe0 0x03
will overlap with the Henkan and Muhenkan keys on Japanese
keyboards. Those keys will geneate the keycode 129 (0x81)
- symbols: XFER and I01 - and 131 (0x83) - symbols NFER and
I03 - in X.
The former keycode however maps to the XF86AudioMedia
keysym on inet keyboards.
From the rules the keysyms are aggregated such that the
mapping of the Henkan key on jp106 keyboards (from the jp
symbol file) gets overwritten by the XF86AudioMedia mapping.
This makes Japanese users loose a key which is rather important
for Japanse input methods only to gain a multimedia key which
is rather irrelevant to people doing serious typing.
This commit adds a huge kludge to fix this: it adds a rule
which in turn adds a statement to the symbol string that
restores the original mapping of this key on Japanse keyboard
layouts. This way Japanse users will still use the XF86AudioMedia
key.
This of course only affects users of the legacy keyboard driver,
evdev doesn't suffer from this problem.
The better solution would probably be to change the keyboard
driver in X to remap the key sequence 0xe0 0x01 to one of the
still unused keycodes in the range above 250.
Then the keyboard mappings can be changed to reflect this.
WIth this change users using an unmatched combination of
xf86-input-keyboard and this package would loose their
XF86AudioMedia key.