Let <KPDL> key on it,fr and nl keyboards generate KP_Separator
Notwithstanding the never ending discussion what character this
key on the numeric keypad should produce this key should produce
what is imprinted on the keyboards. After all all three countries
are observing the SI standards.
This makes this layout more consistent and match for instance
what is done in the de layout.

Signed-off-by: Egbert Eich <eich@freedesktop.org>