symbols/inet: Fix mapping of KEY_DIRECTION/CYCLEWINDOWS for evdev
From checking the usage of KEY_CYCLEWINDOWS in the linux sources, it
seems to be meant for task switching, rather than rotating the screen
as XF86RotateWindows traditionally does. In turn, KEY_DIRECTION is what
some kernel platform modules already use for the "rotate display" key
in certain tablet models.

So in order to make things consistent, it makes more sense to map
KEY_DIRECTION to XF86RotateWindows, and KEY_CYCLEWINDOWS to XF86TaskPane
as it's the closest offered behavior.

Signed-off-by: Carlos Garnacho <carlosg@gnome.org>