Add 'Windows' specific variant to dk, es, fi, it and no layouts
Keyboards for use with Windows seem to have symbols imprinted
on some layouts which don't match the mappings defined by national
standards and/or what has chosen as default for the respective
language.
Here a second EuroSymbol is added to the AltGr-5 key for the
dk, es, fi, it and no layouts.
A language specific variant model has been chosen to be able
to extend these mappings on a per language basis in the future.

Signed-off-by: Egbert Eich <eich@freedesktop.org>