symbols/in: fix unrecognized keysyms
There is no keysym similar to "gur_visarga". Changed according to
http://en.wikipedia.org/wiki/Gurmukhī_alphabet#Visarg :

    The visarg symbol (ਃ U+0A03) is used very occasionally in Gurmukhi. It
    can either represent an abbreviation (like period is used in English) or
    it can act like a Sanskrit Visarga where a voiceless ‘h’ sound is
    pronounced after the vowel.

xkbcommon: WARNING: in:1395:42: unrecognized keysym

There is not "dead_V" keysym and I haven't managed to find what it ought
to be, see:
https://bugs.freedesktop.org/show_bug.cgi?id=19753

Since it has no effect, let's remove it and let the keymap author in the
bug add something useful if desired.

xkbcommon: WARNING: in:1641:46: unrecognized keysym

Signed-off-by: Ran Benita <ran234@gmail.com>