macintosh_vndr/us: include "keypad(overlay)" explicitly
With current xkbcomp, include statements do not obey default annotations
(it does obey them when given a file directly), but libxkbcommon does.
So in xkbcomp, `include "keypad"` includes the first map in the keypad
file, which is "overlay", but includes the default in libxkbcommon,
which is "x11".

"overlay" might in fact be wrong, especially since the keycodes
necessary for the overlay to work are not defined anywhere. But
some users may rely on this in some way, so let's not change it.

Signed-off-by: Ran Benita <ran234@gmail.com>