Define a keyboard layout for the Ol Chiki script
List Santali as a language for the Bangla, Nepali, Ol Chiki and Oriya
keyboard layouts.  According to Ethnologue, Santali gets written in
the following scripts: Bengali [Beng], Devanagari [Deva], Latin [Latn],
Ol Chiki [Olck], Oriya [Orya]. Also according to Ethnologue, the use of
Santali-written-in-Devanagari [sat-Deva] is limited to Nepal, which is
why this change is listing the Santali language only for the Nepali
keyboard, not the Hindi keyboard. http://www.ethnologue.com/language/sat