Make us(dvp) work with regular right Alt key again
This reverts patch f2ccdf6, which was added for issue:

  https://bugs.freedesktop.org/show_bug.cgi?id=64673

This issue is however a duplicate of:

  https://bugs.freedesktop.org/show_bug.cgi?id=24027

and it is the resolution of the latter issue that is correct.

Programmer Dvorak is designed to work with right Alt both as a regular
Alt key and as AltGr. Use of the layout with AltGr should be done
through the (orthogonal) lv3:ralt_switch option instead.