Added Finnish DAS layout
DAS layout is made for Finnish language and it takes account the letter
distribution in Finnish. The layout was invented by Cristian Seres and
the layout in symbols file is originally made by Joel Lehtonen. They
both are mentioned in symbols/fi.

I merely reformatted the layout definition to match other layouts in
symbols/fi and made this patch.

This patch adds DAS layout to symbols/fi and its rule to
rules/base.xml.in.