Move Indian keyboard layouts from base.xml.in to base.extras.xml.in
Apparently the Indian keyboard layouts in xkeyboard-config are not
very useful because of limitations in xkb. In xkb, it is not possible
to output several characters when one key is pressed but that seems
sometimes necessary.

For Indian languages, the input methods supplied by ibus-m17n appear
to be more useful, having the xkb keyboard layouts show up as well in
the gnome-control centre confuse the user because of too many useless
choices.

Therefore, Anish Patil suggested to hide these Indian keyboard layouts
in the gnome-control-centre.

The easiest way to hide them is to move them from
/usr/share/X11/xkb/rules/base.xml to
/usr/share/X11/xkb/rules/base.extras.xml because gnome-control-centre
does not list the layouts from base.extras.xml.

base.extras.xml seems to be for rarely used layouts, so these Indian
layouts seem to fet better there than in base.xml.