symbols/de: Do not specify the key.type where it isn't needed.
With a newer xkbcomp, a trailing "any" (NoSymbol) in a key definition is
taken to mean "none" (VoidSymbol) when the layout specifies the keytype
explicitly.  Avoid this by not specifying the keytype -- it isn't needed
when an included file has already defined the relevant number of levels.

This fixes https://bugs.freedesktop.org/show_bug.cgi?id=97336, a bug
for which a partial fix was applied in commit ae052b1.

Signed-off-by: Benno Schulenberg <bensberg@justemail.net>